(refactor) move config var to independent file
